create chart/requirements.txt
pandas
matplotlib
Pillow
python-dotenv

ipython
google-genai
